+++ TemaLab/quicksort.h
@@ -4,9 +4,9 @@
 
 typedef unsigned long long ULL;
 
-int partition_median3(std::vector<ULL>& nums, int start, int end);
-int partition_middle(std::vector<ULL>& nums, int start, int end);
+size_t partition_median3(std::vector<ULL>& nums, size_t start, size_t end);
+size_t partition_middle(std::vector<ULL>& nums, size_t start, size_t end);
 
-void quick_sort_median3(std::vector<ULL>& nums, int start, int end);
+void quick_sort_median3(std::vector<ULL>& nums, size_t start, size_t end);
 
-void quick_sort_middle(std::vector<ULL>& nums, int start, int end);
+void quick_sort_middle(std::vector<ULL>& nums, size_t start, size_t end);
